+++ backend/src/activation/activation.service.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable, Logger } from '@nestjs/common';
 import { BanCabinetDto } from './dto/ban-cabinet.dto';
 import { InactivatedCabinetDto } from './dto/inactivated-cabinet.dto';
 import { PatchActivationDto } from './dto/patch-activation.dto';
-import { IActivationRepository } from './repository/activation.interface.repository';
+import { IActivationRepository } from './repository/activation.repository.interface';
 
 @Injectable()
 export class ActivationService {
