+++ backend/src/cabinet/repository/cabinet.repository.interface.ts
