+++ backend/src/lent/repository/lent.repository.interface.ts
